Start introducing the notion of method classes
