+++ src/flickr_photos_api/api/__init__.py
@@ -10,14 +10,13 @@ from nitrate.xml import (
     find_required_text,
 )
 
-from .base import HttpxFlickrApi as NewBaseApi
-from ..exceptions import LicenseNotFound
+from .base import HttpxImplementation as NewBaseApi, HttpxImplementation
+from .license_methods import LicenseMethods
 from ..types import (
     CollectionOfElements,
     CollectionOfPhotos,
     Comment,
     GroupInfo,
-    License,
     ParsedElement,
     PhotosFromUrl,
     PhotosInAlbum,
@@ -39,7 +38,15 @@ from ..utils import (
 )
 
 
-class BaseApi(NewBaseApi):
+class FlickrApiMethods(LicenseMethods):
+    pass
+
+
+class FlickrApi(HttpxImplementation, FlickrApiMethods):
+    pass
+
+
+class BaseApi(NewBaseApi, FlickrApiMethods):
     def _to_parsed_element(self, photo_elem: ET.Element) -> ParsedElement:
         """
         Given a <photo> element from a collection response, parse some
@@ -189,75 +196,6 @@ class BaseApi(NewBaseApi):
         # to satisfy coverage.
         else:  # pragma: no cover
             assert False
-
-    @functools.lru_cache()
-    def get_licenses(self) -> dict[str, License]:
-        """
-        Returns a list of licenses, arranged by code.
-
-        See https://www.flickr.com/services/api/flickr.photos.licenses.getInfo.htm
-        """
-        license_resp = self.call(method="flickr.photos.licenses.getInfo")
-
-        result: dict[str, License] = {}
-
-        # Add a short ID which can be used to more easily refer to this
-        # license throughout the codebase.
-        license_ids = {
-            "All Rights Reserved": "in-copyright",
-            "Attribution-NonCommercial-ShareAlike License": "cc-by-nc-sa-2.0",
-            "Attribution-NonCommercial License": "cc-by-nc-2.0",
-            "Attribution-NonCommercial-NoDerivs License": "cc-by-nc-nd-2.0",
-            "Attribution License": "cc-by-2.0",
-            "Attribution-ShareAlike License": "cc-by-sa-2.0",
-            "Attribution-NoDerivs License": "cc-by-nd-2.0",
-            "No known copyright restrictions": "nkcr",
-            "United States Government Work": "usgov",
-            "Public Domain Dedication (CC0)": "cc0-1.0",
-            "Public Domain Mark": "pdm",
-        }
-
-        license_labels = {
-            "Attribution-NonCommercial-ShareAlike License": "CC BY-NC-SA 2.0",
-            "Attribution-NonCommercial License": "CC BY-NC 2.0",
-            "Attribution-NonCommercial-NoDerivs License": "CC BY-NC-ND 2.0",
-            "Attribution License": "CC BY 2.0",
-            "Attribution-ShareAlike License": "CC BY-SA 2.0",
-            "Attribution-NoDerivs License": "CC BY-ND 2.0",
-            "Public Domain Dedication (CC0)": "CC0 1.0",
-        }
-
-        for lic in license_resp.findall(".//license"):
-            result[lic.attrib["id"]] = {
-                "id": license_ids[lic.attrib["name"]],
-                "label": license_labels.get(lic.attrib["name"], lic.attrib["name"]),
-                "url": lic.attrib["url"] or None,
-            }
-
-        return result
-
-    @functools.lru_cache(maxsize=None)
-    def lookup_license_by_id(self, *, id: str) -> License:
-        """
-        Given a license ID from the Flickr API, return the license data.
-
-        e.g. a Flickr API response might include a photo in the following form:
-
-            <photo license="0" …>
-
-        Then you'd call this function to find out what that means:
-
-            >>> api.lookup_license_by_id(id="0")
-            {"id": "in-copyright", "name": "All Rights Reserved", "url": None}
-
-        See https://www.flickr.com/services/api/flickr.photos.licenses.getInfo.htm
-        """
-        licenses = self.get_licenses()
-
-        try:
-            return licenses[id]
-        except KeyError:
-            raise LicenseNotFound(license_id=id)
 
 
 class FlickrPhotosApi(BaseApi):
